requirements.txt
argon2==0.1.10
beautifulsoup4==4.14.3
boto3==1.37.38
cryptography==2.8
faiss_gpu==1.7.2
fasttext==0.9.3
Flask==3.1.2
Flask_Cors==5.0.0
ImageHash==4.3.2
mysql_connector_repackaged==0.3.1
numpy==1.24.4
numpy==1.17.4
openai==2.16.0
Pillow==7.0.0
Pillow==12.1.0
playwright==1.48.0
playwright_stealth==1.0.6
prometheus_client==0.21.1
protobuf==3.6.1
Pygments==2.3.1
redis==5.2.1
Requests==2.32.5
selenium==4.40.0
selenium_stealth==1.0.6
typesense==1.0.3
undetected_chromedriver==3.5.5
webdriver_manager==4.0.2
